Fixed scroll to top component
- fixed scroll to top component in pet and shelter results pages, so that we don't clear the search query
