+++ src/pages/pets.tsx
@@ -7,7 +7,7 @@ import type {
 import { useTranslation } from 'next-i18next';
 import i18nConfig from 'next-i18next.config.mjs';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
-import Link from 'next/link';
+import { useRouter } from 'next/navigation';
 import { useCallback, useEffect, useState } from 'react';
 import AdoptionFormCard from '~/components/cards/AdoptionFormCard';
 import FilterPetsResultsForm from '~/components/forms/FilterPetsResultsForm';
@@ -67,6 +67,7 @@ const SORT_OPTIONS = [
 const Results: NextPage<
   InferGetServerSidePropsType<typeof getServerSideProps>
 > = ({ searchQuery }) => {
+  const router = useRouter();
   const [mobileFiltersOpen, setMobileFiltersOpen] = useState(false);
   const [filter, setFilter] = useState<IPetFilterOptions>({
     sortBy: 'none',
@@ -238,14 +239,13 @@ const Results: NextPage<
           </div>
         )}
         {!hasNextPage && (
-          <Link
-            href={links.scrollToPosition}
-            scroll={false}
+          <span
+            onClick={() => router.push(links.scrollToPosition)}
             className="z-50 mx-auto -mt-6 flex w-fit justify-center"
           >
             <span className="sr-only">Scroll to top</span>
             <Icons.doubleChevronUp className="duration-50 h-12 w-12 cursor-pointer text-primary-300 transition-transform ease-in-out hover:scale-95" />
-          </Link>
+          </span>
         )}
       </main>
     </PageLayout>
