+++ src/server/api/routers/shelter.ts
@@ -5,9 +5,9 @@ import { z } from 'zod';
 import InviteToOrganizationEmail from '~/components/emails/InviteToOrganizationEmail';
 import JoinRequestRejected from '~/components/emails/JoinRequestRejectedEmail';
 import { Roles } from '~/lib/constants';
+import { getSheltersSortOrderBy } from '~/lib/utils';
 import { shelterSettingsSchema } from '~/lib/validators/shelterValidation';
 import { createTRPCRouter } from '~/server/api/trpc';
-import { getSheltersSortOrderBy } from '~/server/helpers/getSortOrderBy';
 import { sheltersSearchWhereConditions } from '~/server/helpers/searchConditions';
 import adminProcedure from '../procedures/adminProcedure';
 import protectedProcedure from '../procedures/protectedProcedure';
